Using SID instead of "display" name doesn't break for Windows ACL groups
Windows systems that have been set up with an image with a base language
that is not English will not have BUILTIN\Administrators but localised
groups instead. Language packs will not fix this, this will stay
forever.

However, Microsoft also uses well-known SIDs, this group among them,
which we can use independent of any localisation (see:
https://learn.microsoft.com/en-us/windows/win32/secauthz/well-known-sids).

This code will fix the issue for non-US systems (thanks for the hint,
Dominik!)